pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "openlist-sdk"
version = "0.1.0"
description = "Python SDK for the OpenList HTTP API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "requests>=2.31.0",
    "PyYAML>=6.0.2",
]

[tool.setuptools]
packages = ["openlist_sdk"]
